dags/comprehend/requirements.txt
boto3==1.16.63
botocore==1.19.63
certifi==2020.12.5
chardet==4.0.0
idna==2.10
jmespath==0.10.0
oauthlib==3.1.0
psycopg2-binary==2.8.6
PySocks==1.7.1
python-dateutil==2.8.1
requests==2.25.1
requests-oauthlib==1.3.0
s3transfer==0.3.4
six==1.15.0
tweepy==3.10.0
urllib3==1.26.3
